Delete Authorization

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController{
 
@@ -15,8 +17,13 @@ class ListingController{
         $this->db = new Database($config);
     }
 
+    /**
+     * Show all listings
+     * 
+     * @return void
+     */
     public function index() {
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
         loadView('listings/index', [
         'listings' => $listings
         ]);
@@ -53,7 +60,7 @@ class ListingController{
     }
 
     /**
-     * Store data in database
+     * Store a listing
      * 
      * @return void
      */
@@ -62,7 +69,7 @@ class ListingController{
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -131,9 +138,16 @@ class ListingController{
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        // Check if listing exists
         if(!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if(!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You are not authorised to delete this listing.';
+            return redirect('/listings/' . $listing->id);
         }
 
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
